model_data_loader/requirements.txt
openpyxl==3.0.10
pandas==1.5.3
python-dotenv==1.0.0
python_dateutil==2.8.2
SQLAlchemy==2.0.8
psycopg2==2.9.3
loguru==0.7.0
